socket-python/requirements.txt
python-socketio~=5.6.0
websocket-client~=1.3.2
inflection
requests